Implement create project
